Update openbee.lua
Updating with static table of bees. This is because FTB Inifnity Evolved has mods which break the API used to find this information programatically. I used a different mod pack that had the API work and printed out the table. A couple other functions have changed to fix other broken APIs